Move WriterArray to separate file
