+++ src/CMakeLists.txt
@@ -157,6 +157,8 @@ set(kafka_to_nexus_INC
   schemas/ev42/ev42_rw.h
   schemas/ev42/ev42_synth.h
   schemas/f142/f142_rw.h
+  schemas/f142/WriterTypedBase.h
+  schemas/f142/WriterArray.h
   schemas/f142/f142_synth.h
   schemas/template/TemplateWriter.h
   CLIOptions.h
